# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: NCV Mail API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /profile:
    get:
      operationId: getProfile
      tags: [profile]
      summary: Get user profile
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"

    patch:
      operationId: updateProfile
      tags: [profile]
      summary: Update profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProfileBody"
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
  /profile/push-token:
    post:
      operationId: registerPushToken
      tags: [profile]
      summary: Register push notification token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterPushTokenBody"
      responses:
        "200":
          description: Token registered
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
...
  /dashboard/category-counts:
    get:
      operationId: getCategoryCounts
      tags: [dashboard]
      summary: Email counts per category
      parameters:
        - in: query
          name: scope
          required: false
          schema:
            type: string
            enum: [personal, shared, all]
          description: Scope du compteur — personal=perso uniquement, shared=boîte partagée (sharedMailboxId requis), all=pooled (défaut)
        - in: query
          name: sharedMailboxId
          required: false
          schema:
            type: string
          description: ID de la boîte partagée quand scope=shared
        - in: query
          name: accountEmail
          required: false
          schema:
            type: string
          description: Filtre par adresse destinataire (scope=personal) — aligne le compteur sur le filtre "compte" de la liste
      responses:
        "200":
          description: Category counts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CategoryCount"
...
components:
  schemas:
...
    UserProfile:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        fullName:
          type: string
        plan:
          type: string
        seats:
          type: integer
        emailsUsed:
          type: integer
        aiCreditsUsed:
          type: integer
          default: 0
        emailsQuota:
          type: integer
        quotaPeriodStart:
          type: ["string", "null"]
          format: date-time
        aiLanguage:
          type: string
          enum: [fr, en, nl]
          default: fr
        signature:
          type: string
          default: ""
        timezone:
          type: string
          default: "Europe/Brussels"
        createdAt:
          type: string
          format: date-time
        organisationId:
          type: ["string", "null"]
        organisationName:
          type: ["string", "null"]
        organisationRole:
          type: ["string", "null"]
          enum: [admin, member, null]
        isAdmin:
          type: boolean
          default: false
          description: True when the profile has the internal admin flag.
        followUpDelayDays:
          type: integer
          default: 5
          minimum: 1
          maximum: 60
          description: Délai par défaut (en jours) avant de suggérer une relance sur un mail envoyé sans réponse.
        trackingEnabled:
          type: boolean
          default: false
          description: True if the user opted in to read-receipt tracking pixels on sent emails.
      required:
        - id
        - email
        - fullName
        - plan
        - seats
        - emailsUsed
        - emailsQuota
        - createdAt
    UpdateProfileBody:
      type: object
      properties:
        fullName:
          type: string
        plan:
          type: string
        seats:
          type: integer
        aiLanguage:
          type: string
          enum: [fr, en, nl]
        signature:
          type: string
        timezone:
          type: string
        followUpDelayDays:
          type: integer
          minimum: 1
          maximum: 60
        trackingEnabled:
          type: boolean
    RegisterPushTokenBody:
      type: object
      required:
        - token
      properties:
        token:
          type: string
        platform:
          type: string
...
    CategoryCount:
      type: object
      properties:
        categoryId:
          type: integer
        categoryName:
          type: string
        count:
          type: integer
      required:
        - categoryId
        - categoryName
        - count
...