# lib/api-zod/src/generated/types/updateProfileBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * NCV Mail API specification
 * OpenAPI spec version: 0.1.0
 */
import type { UpdateProfileBodyAiLanguage } from "./updateProfileBodyAiLanguage";

export interface UpdateProfileBody {
  fullName?: string;
  plan?: string;
  seats?: number;
  aiLanguage?: UpdateProfileBodyAiLanguage;
  signature?: string;
  timezone?: string;
  /**
   * @minimum 1
   * @maximum 60
   */
  followUpDelayDays?: number;
  trackingEnabled?: boolean;
}