# lib/api-zod/src/generated/types/userProfile.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * NCV Mail API specification
 * OpenAPI spec version: 0.1.0
 */
import type { UserProfileAiLanguage } from "./userProfileAiLanguage";
import type { UserProfileOrganisationRole } from "./userProfileOrganisationRole";

export interface UserProfile {
  id: number;
  email: string;
  fullName: string;
  plan: string;
  seats: number;
  emailsUsed: number;
  aiCreditsUsed?: number;
  emailsQuota: number;
  /** @nullable */
  quotaPeriodStart?: Date | null;
  aiLanguage?: UserProfileAiLanguage;
  signature?: string;
  timezone?: string;
  createdAt: Date;
  /** @nullable */
  organisationId?: string | null;
  /** @nullable */
  organisationName?: string | null;
  /** @nullable */
  organisationRole?: UserProfileOrganisationRole;
  /** True when the profile has the internal admin flag. */
  isAdmin?: boolean;
  /**
   * Délai par défaut (en jours) avant de suggérer une relance sur un mail envoyé sans réponse.
   * @minimum 1
   * @maximum 60
   */
  followUpDelayDays?: number;
  /** True if the user opted in to read-receipt tracking pixels on sent emails. */
  trackingEnabled?: boolean;
}